% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsicons-package.R
\docType{package}
\name{bsicons-package}
\alias{bsicons}
\alias{bsicons-package}
\title{bsicons: Easily Work with 'Bootstrap' Icons}
\description{
Easily use 'Bootstrap' icons inside 'Shiny' apps and 'R Markdown' documents. More generally, icons can be inserted in any 'htmltools' document through inline 'SVG'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/bsicons}
  \item Report bugs at \url{https://github.com/rstudio/bsicons/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{carson@posit.co} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Mark Otto (Bootstrap icons maintainer) [copyright holder]
}

}
\keyword{internal}
